requirements.txt
beautifulsoup4==4.8.1
colorama==0.3.7
gensim==3.8.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
graphviz==0.13
html5lib==0.999999999
httplib2==0.14.0
hyperlink==17.3.1
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.2.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
nltk==3.4.5
notebook==6.0.2
numexpr==2.7.0
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
pickleshare==0.7.5
py==1.8.0
pyLDAvis==2.1.2
pyOpenSSL==17.5.0
pyserial==3.4
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
scikit-learn==0.21.3
scipy==1.3.3
sklearn==0.0
soupsieve==1.9.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
torchviz==0.0.1
urllib3==1.25.7
wikipedia==1.4.0
cycler-0.10.0 
kiwisolver-1.1.0 
pillow-6.2.1 
pyparsing-2.4.5 
python-dateutil-2.8.1 
setuptools-42.0.2 
six-1.13.0 
wordcloud-1.6.0    